+++ CHANGES.md
@@ -1,6 +1,8 @@
 ### unreleased
 
 #### Yocaml
+
+- Add `Data.Validation.where_opt` (and `String`, `Int` and `Float` version) (by [xvw](https://xvw.lol))
 - Improve pretty-printing of validation errors (by [Linda-Njau](https://github.com/Linda-Njau))
 - Fix typos and improve logs display (by [clementd](https://clementd.wtf))